minor bugfix and more pysical aproche
fixed bug with warning, with double condition for single Mask setting

switched from calculating the derivative of binning data, to actual
binning of the data.
This is essentialy the same but mor physical and without the need of
finite differences.
This changes the result slightly but the cutoff is roughly the same
without noticable consequences.